Default to bf16, and record why B=32 S=8192 measures faster than B=1 S=16384
The shape question first, because the answer is not the one it looks like. Both
shapes do exactly the same work -- 8.796 TFLOP per dispatch -- and both divide
evenly into 256 resident workgroups, so the 11% gap is neither a FLOP-count nor a
tail artifact. The loop body is identical: in-loop MFMA efficiency and cyc/iter
barely move (fav4 93.9 -> 94.5%, 4361.5 -> 4334.6), which they cannot, since the
loop does not depend on B or H.

Per cycle the B=32 shape is strictly worse. Each workgroup walks 128 K/V blocks
instead of 256, so the fixed prologue and drain amortize over half as much work
and their share doubles, 5.9% -> 11.7%, which is about 4% more cycles per
dispatch. It wins anyway, on clock, because it is not power-limited: fav4 at
B=1 S=16384 sits at 1396.5 W, pinned to the ~1400 W cap at 1527.7 MHz, while at
B=32 S=8192 it draws 1317.6 W -- 79 W below the cap -- and runs at 1569.8 MHz.
A less MFMA-dense stream draws less power, and off the cap that converts into
frequency worth more than the cycles it cost. The better config is better partly
because it wastes more time on pipeline fill and drain.

That also explains why the ranking flips between shapes, which matters for how
section 8 should be read. The two kernels have opposite strengths: fav4 is
cycle-efficient (94.5% in-loop against FlyDSL's 84.7%) and FlyDSL is
power-efficient per cycle (at B=32 it clocks 5.2% higher at comparable power, and
spends more of its time in the loop -- 94.13% against our 88.33%, so its pipeline
fill and drain are cheaper than our four-stage one). Where the board is pinned,
cycles decide and fav4 wins by 9.3%; where there is headroom the two cancel:

  fav4 cycle advantage   (0.945 x 0.8833) / (0.847 x 0.9413) = +4.8%
  FlyDSL clock advantage  1650.9 / 1569.8                    = +5.2%

and the measurement agrees, 1318.0 against 1320.1. Neither shape gives "the"
answer -- B=1 S=16384 flatters us and B=32 S=8192 flatters them.

The B=32 matrix is recorded with it (bf16, interleaved, three rounds). FlyDSL's
published claim of 1320 TFLOPS at this config reproduces at 1319.4 / 1319.6 /
1320.2. Every row is stable to about 2 TFLOPS here, against a 34-point spread and
a 7% warm-up decay at B=1 S=16320, because 8192 workgroups and a 2.1 GB working
set reach a uniform steady state -- so this shape is the better one for headline
numbers, for both projects.

bf16 becomes the default for bench.py, fa_kernel_time.py and fly_kernel_time.py.
It is what these parts are usually run in, and it measures 3-7% faster than fp16
at identical cycle counts and identical in-loop efficiency -- the whole difference
is clock, since bf16's narrower mantissa toggles less of the multiplier array. The
documented defaults in section 9 are updated to match.

+++ kernels/attention/bench.py
@@ -102,7 +102,7 @@ SEQLENS = [1024, 2048, 4096, 8192, 16384]
 
 def parse_args():
     p = argparse.ArgumentParser(description="FAV3 rotated-4cluster attention benchmark (gfx950)")
-    p.add_argument("--dtype", choices=["fp16", "bf16"], default="fp16")
+    p.add_argument("--dtype", choices=["fp16", "bf16"], default="bf16")
     p.add_argument("--layout", choices=["bhsd", "bshd"], default="bhsd")
     p.add_argument("--batch", type=int, default=1)
     p.add_argument("--hq", type=int, default=64, help="number of query heads")
